benchmarks/MNIST_FCNeuralNetwork/requirements.txt
tensorflow-gpu
matplotlib
sklearn
gputil
numpy==1.21
numba
nvidia-ml-py3